requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyMuPDF==1.23.6
pdf2docx==0.5.6
pdfplumber==0.10.3
pandas==2.1.3
openpyxl==3.1.2
Pillow==10.1.0